requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
tensorflow>=2.15.0,<2.16.0
keras>=2.15.0,<2.16.0
keras-preprocessing>=1.1.2
librosa>=0.10.1
soundfile>=0.12.1
opencv-python>=4.8.0
mtcnn>=0.1.1
pandas>=2.1.0
numpy>=1.24.0,<2.0.0
yt-dlp>=2023.12.30
google-genai>=0.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0

